Updated main.py to run with Team 3's part, Also change the file that runs both the frontend and backend to 'run.py'
