--- Backend/Team 3/Database/test.txt
@@ -1,2 +1,0 @@
-Added to commit directories
-delete later
