+++ Backend/Team3/Database/Analyst_Manager.py
