+++ Backend/Team3/Database/Database_Manager.py
